+++ wpull/http.py
@@ -53,7 +53,7 @@ class Request(BaseRequest):
         self.url_info = None
         self.version = version
         self.fields = NameValueRecord()
-        self.body = None
+        self.body = Body()
         self.address = None
 
     @classmethod
@@ -105,6 +105,8 @@ class Response(BaseResponse):
             :class:`.namevalue.NameValueRecord` containing the HTTP header
             (and trailer, if present) fields.
         body (Body): An instance of :class:`.conversation.Body`.
+        url_info (URLInfo): An instance of :class:`.url.URLInfo` for the
+            of the corresponding request.
     '''
     def __init__(self, version, status_code, status_reason):
         self.version = version
@@ -112,6 +114,7 @@ class Response(BaseResponse):
         self.status_reason = status_reason
         self.fields = NameValueRecord()
         self.body = Body()
+        self.url_info = None
 
     @classmethod
     def parse_status_line(cls, string):
@@ -156,6 +159,7 @@ class Response(BaseResponse):
             'status_code': self.status_code,
             'status_reason': self.status_reason,
             'body': self.body.to_dict(),
+            'url_info': self.url_info.to_dict() if self.url_info else None
         }
 
 
@@ -317,6 +321,7 @@ class Connection(object):
             else:
                 response = yield self._process_request(request,
                     response_factory)
+                response.url_info = request.url_info
         except:
             _logger.debug('Fetch exception.')
             self.close()
@@ -767,6 +772,69 @@ class Client(BaseClient):
 
         if self._recorder:
             self._recorder.close()
+
+
+class RedirectTracker(object):
+    '''Keeps track of HTTP document URL redirects.
+
+    Args:
+        max_redirects (int): The maximum number of redirects to allow.
+        codes: The HTTP status codes indicating a redirect where the method
+            can change to "GET".
+        repeat_codes: The HTTP status codes indicating a redirect where
+            the method cannot change and future requests should be repeated.
+    '''
+    REDIRECT_CODES = (301, 302, 303)
+    REPEAT_REDIRECT_CODES = (307, 308)
+
+    def __init__(self, max_redirects=20, codes=REDIRECT_CODES,
+    repeat_codes=REPEAT_REDIRECT_CODES):
+        self._max_redirects = max_redirects
+        self._codes = codes
+        self._repeat_codes = repeat_codes
+        self._response = None
+        self._num_redirects = 0
+
+    def load(self, response):
+        '''Load the response and increment the counter.
+
+        Args:
+            response (Response): An instance of :class:`Response`.
+        '''
+        self._response = response
+
+        if self.next_location():
+            self._num_redirects += 1
+
+    def next_location(self):
+        '''Returns the next location.
+
+        Returns:
+            str, None: If str, the raw string contained in the Location field.
+            Otherwise, no next location.
+        '''
+        if self._response:
+            return self._response.fields.get('location')
+
+    def is_redirect(self):
+        '''Return whether the response contains a redirect code.'''
+        if self._response:
+            status_code = self._response.status_code
+            return status_code in self._codes \
+                or status_code in self._repeat_codes
+
+    def is_repeat(self):
+        '''Return whether the next request should be repeated.'''
+        if self._response:
+            return self._response.status_code in self._repeat_codes
+
+    def count(self):
+        '''Return the number of redirects received so far.'''
+        return self._num_redirects
+
+    def exceeded(self):
+        '''Return whether the number of redirects has exceeded the maximum.'''
+        return self._num_redirects > self._max_redirects
 
 
 def parse_charset(header_string):