+++ doc/changelog.rst
@@ -2,6 +2,17 @@
 What's New
 ==========
 
+0.14 (2014-02-01)
+=================
+
+* Database change: SQLAlchemy is now used for the URL Table.
+
+  * Scripts: ``url_info['inline']`` now returns a boolean, not an integer.
+
+* Implements ``--post-data`` and ``--post-file``.
+* Scripts can now return ``post_data`` and ``link_type`` as part of ``get_urls()``.
+
+
 0.13 (2014-01-31)
 =================